Completed the Controller implementation
Completed a mecanum drive base implementation in DriveSubsystem. This does not have working odometry code as of now, which means field centric code does not work. There is space specifically for the odometry code to be updated.

Added an example ShooterSubsystem, with example commands. Command Scheduler is not fully tested yet, which is what this is for.

Fitted a RobotContainer Class for these implementations. Cannot still figure out why telemetry is being problematic when it leaves the opmode class.